+++ src/Symfony/Component/Messenger/Event/AbstractWorkerMessageEvent.php
@@ -11,13 +11,12 @@
 
 namespace Symfony\Component\Messenger\Event;
 
-use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\Messenger\Envelope;
 
 /**
  * @experimental in 4.3
  */
-abstract class AbstractWorkerMessageEvent extends Event
+abstract class AbstractWorkerMessageEvent
 {
     private $envelope;
     private $receiverName;