+++ src/Symfony/Component/Security/Core/Authentication/AuthenticationProviderManager.php
@@ -51,6 +51,9 @@ class AuthenticationProviderManager implements AuthenticationManagerInterface
         $this->eraseCredentials = $eraseCredentials;
     }
 
+    /**
+     * @final since Symfony 4.3, the type-hint will be updated to the interface from symfony/contracts in 5.0
+     */
     public function setEventDispatcher(EventDispatcherInterface $dispatcher)
     {
         $this->eventDispatcher = LegacyEventDispatcherProxy::decorate($dispatcher);