+++ src/Symfony/Component/Security/Http/Firewall/ExceptionListener.php
@@ -39,7 +39,7 @@ use Symfony\Component\Security\Http\Util\TargetPathTrait;
  *
  * @author Fabien Potencier <fabien@symfony.com>
  *
- * @final since Symfony 4.3
+ * @final since Symfony 4.3, EventDispatcherInterface type-hints will be updated to the interface from symfony/contracts in 5.0
  */
 class ExceptionListener
 {